requirements.txt
# ── Core API ──────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# ── PDF Processing ────────────────────────────────────────────────────────────
pymupdf==1.24.3          # fast PDF text extraction (fitz)

# ── NLP / Embeddings ──────────────────────────────────────────────────────────
sentence-transformers==2.7.0   # local embeddings (all-MiniLM-L6-v2)
langchain==0.1.20
langchain-community==0.0.38
langchain-text-splitters==0.0.2

# ── Vector Store ──────────────────────────────────────────────────────────────
chromadb==0.5.0          # persistent local vector DB
faiss-cpu==1.8.0         # fast similarity search

# ── LLM API ───────────────────────────────────────────────────────────────────
groq==0.6.0              # Groq API (free, blazing fast llama3)

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
aiofiles==23.2.1
numpy==1.26.4
tiktoken==0.7.0          # token counting

# ── Dev ───────────────────────────────────────────────────────────────────────
pytest==8.2.0
httpx==0.27.0            # test FastAPI
